requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
Django==5.1.1
django-allauth==65.0.2
django-allauth-themes==0.1.2
django-browser-reload==1.15.0
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-grappelli==4.0.1
django-jazzmin==3.0.1
idna==3.10
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.3
whitenoise==6.8.2
